ch4/ofi: split ofi_init.c into smaller units
ofi_init.c is too large and contains quite a few components that each
with its own set of static functions. Split them to facilitate
development.
