+++ src/mpid/ch4/netmod/ofi/Makefile.mk
@@ -8,6 +8,9 @@ if BUILD_CH4_NETMOD_OFI
 noinst_HEADERS     +=
 mpi_core_sources   += src/mpid/ch4/netmod/ofi/func_table.c \
                       src/mpid/ch4/netmod/ofi/ofi_init.c \
+                      src/mpid/ch4/netmod/ofi/ofi_finalize.c \
+                      src/mpid/ch4/netmod/ofi/ofi_vni.c \
+                      src/mpid/ch4/netmod/ofi/ofi_dynamic.c \
                       src/mpid/ch4/netmod/ofi/ofi_comm.c \
                       src/mpid/ch4/netmod/ofi/ofi_datatype.c \
                       src/mpid/ch4/netmod/ofi/ofi_op.c \
@@ -17,6 +20,7 @@ mpi_core_sources   += src/mpid/ch4/netmod/ofi/func_table.c \
                       src/mpid/ch4/netmod/ofi/ofi_progress.c \
                       src/mpid/ch4/netmod/ofi/globals.c \
                       src/mpid/ch4/netmod/ofi/util.c
+
 errnames_txt_files += src/mpid/ch4/netmod/ofi/errnames.txt
 external_subdirs   += @ofisrcdir@
 pmpi_convenience_libs += @ofilib@
