+++ src/mpid/ch4/netmod/ofi/ofi_init.c
@@ -306,12 +306,6 @@ cvars:
 
 static int get_ofi_version(void);
 static int open_fabric(void);
-static int create_vni_context(int vni);
-static int destroy_vni_context(int vni);
-
-static int conn_manager_init(void);
-static int conn_manager_destroy(void);
-static int dynproc_send_disconnect(int conn_id);
 
 static int get_ofi_version(void)
 {
@@ -319,183 +313,6 @@ static int get_ofi_version(void)
         return FI_VERSION(MPIDI_OFI_MAJOR_VERSION, MPIDI_OFI_MINOR_VERSION);
     else
         return FI_VERSION(FI_MAJOR_VERSION, FI_MINOR_VERSION);
-}
-
-static int conn_manager_init()
-{
-    int mpi_errno = MPI_SUCCESS, i;
-
-    MPIR_FUNC_VERBOSE_STATE_DECL(MPID_STATE_CONN_MANAGER_INIT);
-    MPIR_FUNC_VERBOSE_ENTER(MPID_STATE_CONN_MANAGER_INIT);
-
-    MPIDI_OFI_global.conn_mgr.mmapped_size = 8 * 4 * 1024;
-    MPIDI_OFI_global.conn_mgr.max_n_conn = 1;
-    MPIDI_OFI_global.conn_mgr.next_conn_id = 0;
-    MPIDI_OFI_global.conn_mgr.n_conn = 0;
-
-    MPIDI_OFI_global.conn_mgr.conn_list =
-        (MPIDI_OFI_conn_t *) MPL_mmap(NULL, MPIDI_OFI_global.conn_mgr.mmapped_size,
-                                      PROT_READ | PROT_WRITE, MAP_PRIVATE | MAP_ANON, -1, 0,
-                                      MPL_MEM_ADDRESS);
-    MPIR_ERR_CHKANDSTMT(MPIDI_OFI_global.conn_mgr.conn_list == MAP_FAILED, mpi_errno,
-                        MPI_ERR_NO_MEM, goto fn_fail, "**nomem");
-
-    MPIDI_OFI_global.conn_mgr.free_conn_id =
-        (int *) MPL_malloc(MPIDI_OFI_global.conn_mgr.max_n_conn * sizeof(int), MPL_MEM_ADDRESS);
-    MPIR_ERR_CHKANDSTMT(MPIDI_OFI_global.conn_mgr.free_conn_id == NULL, mpi_errno,
-                        MPI_ERR_NO_MEM, goto fn_fail, "**nomem");
-
-    for (i = 0; i < MPIDI_OFI_global.conn_mgr.max_n_conn; ++i) {
-        MPIDI_OFI_global.conn_mgr.free_conn_id[i] = i + 1;
-    }
-    MPIDI_OFI_global.conn_mgr.free_conn_id[MPIDI_OFI_global.conn_mgr.max_n_conn - 1] = -1;
-
-  fn_exit:
-    MPIR_FUNC_VERBOSE_EXIT(MPID_STATE_CONN_MANAGER_INIT);
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-static int conn_manager_destroy()
-{
-    int mpi_errno = MPI_SUCCESS, i, j;
-    MPIDI_OFI_dynamic_process_request_t *req;
-    fi_addr_t *conn;
-    int max_n_conn = MPIDI_OFI_global.conn_mgr.max_n_conn;
-    int *close_msg;
-    uint64_t match_bits = 0;
-    uint64_t mask_bits = 0;
-    MPIR_Context_id_t context_id = 0xF000;
-    MPIR_CHKLMEM_DECL(3);
-
-    MPIR_FUNC_VERBOSE_STATE_DECL(MPID_STATE_CONN_MANAGER_DESTROY);
-    MPIR_FUNC_VERBOSE_ENTER(MPID_STATE_CONN_MANAGER_DESTROY);
-
-    match_bits = MPIDI_OFI_init_recvtag(&mask_bits, context_id, 1);
-    match_bits |= MPIDI_OFI_DYNPROC_SEND;
-
-    if (max_n_conn > 0) {
-        /* try wait/close connections */
-        MPIR_CHKLMEM_MALLOC(req, MPIDI_OFI_dynamic_process_request_t *,
-                            max_n_conn * sizeof(MPIDI_OFI_dynamic_process_request_t), mpi_errno,
-                            "req", MPL_MEM_BUFFER);
-        MPIR_CHKLMEM_MALLOC(conn, fi_addr_t *, max_n_conn * sizeof(fi_addr_t), mpi_errno, "conn",
-                            MPL_MEM_BUFFER);
-        MPIR_CHKLMEM_MALLOC(close_msg, int *, max_n_conn * sizeof(int), mpi_errno, "int",
-                            MPL_MEM_BUFFER);
-
-        j = 0;
-        for (i = 0; i < max_n_conn; ++i) {
-            switch (MPIDI_OFI_global.conn_mgr.conn_list[i].state) {
-                case MPIDI_OFI_DYNPROC_CONNECTED_CHILD:
-                    mpi_errno = dynproc_send_disconnect(i);
-                    MPIR_ERR_CHECK(mpi_errno);
-                    break;
-                case MPIDI_OFI_DYNPROC_LOCAL_DISCONNECTED_PARENT:
-                case MPIDI_OFI_DYNPROC_CONNECTED_PARENT:
-                    MPL_DBG_MSG_FMT(MPIDI_CH4_DBG_GENERAL, VERBOSE,
-                                    (MPL_DBG_FDEST, "Wait for close of conn_id=%d", i));
-                    conn[j] = MPIDI_OFI_global.conn_mgr.conn_list[i].dest;
-                    req[j].done = 0;
-                    req[j].event_id = MPIDI_OFI_EVENT_DYNPROC_DONE;
-                    MPIDI_OFI_CALL_RETRY(fi_trecv(MPIDI_OFI_global.ctx[0].rx,
-                                                  &close_msg[j],
-                                                  sizeof(int),
-                                                  NULL,
-                                                  conn[j],
-                                                  match_bits,
-                                                  mask_bits, &req[j].context), trecv, FALSE);
-                    j++;
-                    break;
-                default:
-                    break;
-            }
-        }
-
-        for (i = 0; i < j; ++i) {
-            MPIDI_OFI_PROGRESS_WHILE(!req[i].done);
-            MPIDI_OFI_global.conn_mgr.conn_list[i].state = MPIDI_OFI_DYNPROC_DISCONNECTED;
-            MPL_DBG_MSG_FMT(MPIDI_CH4_DBG_GENERAL, VERBOSE,
-                            (MPL_DBG_FDEST, "conn_id=%d closed", i));
-        }
-
-        MPIR_CHKLMEM_FREEALL();
-    }
-
-    MPL_munmap((void *) MPIDI_OFI_global.conn_mgr.conn_list, MPIDI_OFI_global.conn_mgr.mmapped_size,
-               MPL_MEM_ADDRESS);
-    MPL_free(MPIDI_OFI_global.conn_mgr.free_conn_id);
-
-  fn_exit:
-    MPIR_FUNC_VERBOSE_EXIT(MPID_STATE_CONN_MANAGER_DESTROY);
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-static int dynproc_send_disconnect(int conn_id)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    MPIR_Context_id_t context_id = 0xF000;
-    MPIDI_OFI_dynamic_process_request_t req;
-    uint64_t match_bits = 0;
-    unsigned int close_msg = 0xcccccccc;
-    struct fi_msg_tagged msg;
-    struct iovec msg_iov;
-
-    MPIR_FUNC_VERBOSE_STATE_DECL(MPID_STATE_DYNPROC_SEND_DISCONNECT);
-    MPIR_FUNC_VERBOSE_ENTER(MPID_STATE_DYNPROC_SEND_DISCONNECT);
-
-    if (MPIDI_OFI_global.conn_mgr.conn_list[conn_id].state == MPIDI_OFI_DYNPROC_CONNECTED_CHILD) {
-        MPL_DBG_MSG_FMT(MPIDI_CH4_DBG_GENERAL, VERBOSE,
-                        (MPL_DBG_FDEST, " send disconnect msg conn_id=%d from child side",
-                         conn_id));
-        match_bits = MPIDI_OFI_init_sendtag(context_id, 1, MPIDI_OFI_DYNPROC_SEND);
-
-        /* fi_av_map here is not quite right for some providers */
-        /* we need to get this connection from the sockname     */
-        req.done = 0;
-        req.event_id = MPIDI_OFI_EVENT_DYNPROC_DONE;
-        msg_iov.iov_base = &close_msg;
-        msg_iov.iov_len = sizeof(close_msg);
-        msg.msg_iov = &msg_iov;
-        msg.desc = NULL;
-        msg.iov_count = 0;
-        msg.addr = MPIDI_OFI_global.conn_mgr.conn_list[conn_id].dest;
-        msg.tag = match_bits;
-        msg.ignore = context_id;
-        msg.context = (void *) &req.context;
-        msg.data = 0;
-        MPIDI_OFI_CALL_RETRY(fi_tsendmsg(MPIDI_OFI_global.ctx[0].tx, &msg,
-                                         FI_COMPLETION | FI_TRANSMIT_COMPLETE | FI_REMOTE_CQ_DATA),
-                             tsendmsg, FALSE);
-        MPIDI_OFI_PROGRESS_WHILE(!req.done);
-    }
-
-    switch (MPIDI_OFI_global.conn_mgr.conn_list[conn_id].state) {
-        case MPIDI_OFI_DYNPROC_CONNECTED_CHILD:
-            MPIDI_OFI_global.conn_mgr.conn_list[conn_id].state =
-                MPIDI_OFI_DYNPROC_LOCAL_DISCONNECTED_CHILD;
-            break;
-        case MPIDI_OFI_DYNPROC_CONNECTED_PARENT:
-            MPIDI_OFI_global.conn_mgr.conn_list[conn_id].state =
-                MPIDI_OFI_DYNPROC_LOCAL_DISCONNECTED_PARENT;
-            break;
-        default:
-            break;
-    }
-
-    MPL_DBG_MSG_FMT(MPIDI_CH4_DBG_GENERAL, VERBOSE,
-                    (MPL_DBG_FDEST, " local_disconnected conn_id=%d state=%d",
-                     conn_id, MPIDI_OFI_global.conn_mgr.conn_list[conn_id].state));
-
-  fn_exit:
-    MPIR_FUNC_VERBOSE_EXIT(MPID_STATE_DYNPROC_SEND_DISCONNECT);
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
 }
 
 int MPIDI_OFI_mpi_init_hook(int rank, int size, int appnum, int *tag_bits, MPIR_Comm * init_comm)
@@ -561,17 +378,8 @@ int MPIDI_OFI_mpi_init_hook(int rank, int size, int appnum, int *tag_bits, MPIR_
 
     MPIDI_OFI_global.num_ctx = num_vnis;
 
-    /* Create MPIDI_OFI_global.ctx[0] first  */
-    mpi_errno = create_vni_context(0);
+    mpi_errno = MPIDI_OFI_vni_init();
     MPIR_ERR_CHECK(mpi_errno);
-
-    /* Creating the additional vni contexts.
-     * This code maybe moved to a later stage */
-    for (i = 1; i < MPIDI_OFI_global.num_ctx; i++) {
-        mpi_errno = create_vni_context(i);
-        MPIR_ERR_CHECK(mpi_errno);
-    }
-
     /* ------------------------------------------------------------------------ */
     /* Address exchange (essentially activating the vnis)                       */
     /* ------------------------------------------------------------------------ */
@@ -698,10 +506,8 @@ int MPIDI_OFI_mpi_init_hook(int rank, int size, int appnum, int *tag_bits, MPIR_
     /* Initalize RMA keys allocator */
     MPIDI_OFI_mr_key_allocator_init();
 
-    /* ------------------------------------------------- */
     /* Initialize Connection Manager for Dynamic Tasking */
-    /* ------------------------------------------------- */
-    conn_manager_init();
+    MPIDI_OFI_conn_manager_init();
 
     MPIR_Comm_register_hint(MPIR_COMM_HINT_EAGAIN, "eagain", NULL, MPIR_COMM_HINT_TYPE_BOOL, 0);
 
@@ -712,91 +518,6 @@ int MPIDI_OFI_mpi_init_hook(int rank, int size, int appnum, int *tag_bits, MPIR_
     *tag_bits = MPIDI_OFI_TAG_BITS;
 
     MPIR_FUNC_VERBOSE_EXIT(MPID_STATE_MPIDI_OFI_MPI_INIT_HOOK);
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-int MPIDI_OFI_mpi_finalize_hook(void)
-{
-    int mpi_errno = MPI_SUCCESS;
-    int i = 0;
-    int barrier[2] = { 0 };
-    MPIR_Errflag_t errflag = MPIR_ERR_NONE;
-
-    MPIR_FUNC_VERBOSE_STATE_DECL(MPID_STATE_MPIDI_OFI_MPI_FINALIZE_HOOK);
-    MPIR_FUNC_VERBOSE_ENTER(MPID_STATE_MPIDI_OFI_MPI_FINALIZE_HOOK);
-
-    /* clean dynamic process connections */
-    mpi_errno = conn_manager_destroy();
-    MPIR_ERR_CHECK(mpi_errno);
-
-    /* Progress until we drain all inflight RMA send long buffers */
-    while (MPL_atomic_load_int(&MPIDI_OFI_global.am_inflight_rma_send_mrs) > 0)
-        MPIDI_OFI_PROGRESS();
-
-    /* Destroy RMA key allocator */
-    MPIDI_OFI_mr_key_allocator_destroy();
-
-    /* Barrier over allreduce, but force non-immediate send */
-    MPIDI_OFI_global.max_buffered_send = 0;
-    mpi_errno = MPIR_Allreduce_allcomm_auto(&barrier[0], &barrier[1], 1, MPI_INT, MPI_SUM,
-                                            MPIR_Process.comm_world, &errflag);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    /* Progress until we drain all inflight injection emulation requests */
-    while (MPL_atomic_load_int(&MPIDI_OFI_global.am_inflight_inject_emus) > 0)
-        MPIDI_OFI_PROGRESS();
-    MPIR_Assert(MPL_atomic_load_int(&MPIDI_OFI_global.am_inflight_inject_emus) == 0);
-
-    /* Tearing down endpoints */
-    for (i = 1; i < MPIDI_OFI_global.num_ctx; i++) {
-        mpi_errno = destroy_vni_context(i);
-        MPIR_ERR_CHECK(mpi_errno);
-    }
-    /* 0th ctx is special, synonymous to global context */
-    mpi_errno = destroy_vni_context(0);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.fabric->fid), fabricclose);
-
-    fi_freeinfo(MPIDI_OFI_global.prov_use);
-
-    MPIDIU_map_destroy(MPIDI_OFI_global.win_map);
-
-    if (MPIDI_OFI_ENABLE_AM) {
-        while (MPIDI_OFI_global.am_unordered_msgs) {
-            MPIDI_OFI_am_unordered_msg_t *uo_msg = MPIDI_OFI_global.am_unordered_msgs;
-            DL_DELETE(MPIDI_OFI_global.am_unordered_msgs, uo_msg);
-        }
-        MPIDIU_map_destroy(MPIDI_OFI_global.am_send_seq_tracker);
-        MPIDIU_map_destroy(MPIDI_OFI_global.am_recv_seq_tracker);
-
-        for (i = 0; i < MPIDI_OFI_NUM_AM_BUFFERS; i++)
-            MPL_free(MPIDI_OFI_global.am_bufs[i]);
-
-        MPIDIU_destroy_buf_pool(MPIDI_OFI_global.am_buf_pool);
-
-        MPIR_Assert(MPIDI_OFI_global.cq_buffered_static_head ==
-                    MPIDI_OFI_global.cq_buffered_static_tail);
-        MPIR_Assert(NULL == MPIDI_OFI_global.cq_buffered_dynamic_head);
-    }
-
-    int err;
-    MPID_Thread_mutex_destroy(&MPIDI_OFI_THREAD_UTIL_MUTEX, &err);
-    MPIR_Assert(err == 0);
-
-    MPID_Thread_mutex_destroy(&MPIDI_OFI_THREAD_PROGRESS_MUTEX, &err);
-    MPIR_Assert(err == 0);
-
-    MPID_Thread_mutex_destroy(&MPIDI_OFI_THREAD_FI_MUTEX, &err);
-    MPIR_Assert(err == 0);
-
-    MPID_Thread_mutex_destroy(&MPIDI_OFI_THREAD_SPAWN_MUTEX, &err);
-    MPIR_Assert(err == 0);
-
-  fn_exit:
-    MPIR_FUNC_VERBOSE_EXIT(MPID_STATE_MPIDI_OFI_MPI_FINALIZE_HOOK);
     return mpi_errno;
   fn_fail:
     goto fn_exit;
@@ -944,147 +665,6 @@ int MPIDI_OFI_create_intercomm_from_lpids(MPIR_Comm * newcomm_ptr, int size, con
     return 0;
 }
 
-/* ---- static functions for vni contexts ---- */
-static int create_vni_domain(struct fid_domain **p_domain, struct fid_av **p_av,
-                             struct fid_cntr **p_cntr);
-static int create_cq(struct fid_domain *domain, struct fid_cq **p_cq);
-static int create_sep_tx(struct fid_ep *ep, int idx, struct fid_ep **p_tx,
-                         struct fid_cq *cq, struct fid_cntr *cntr);
-static int create_sep_rx(struct fid_ep *ep, int idx, struct fid_ep **p_rx,
-                         struct fid_cq *cq, struct fid_cntr *cntr);
-static int try_open_shared_av(struct fid_domain *domain, struct fid_av **p_av);
-static int create_rma_stx_ctx(struct fid_domain *domain, struct fid_stx **p_rma_stx_ctx);
-
-static int create_vni_context(int vni)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    MPIR_FUNC_VERBOSE_STATE_DECL(MPID_STATE_CREATE_VNI_CONTEXT);
-    MPIR_FUNC_VERBOSE_ENTER(MPID_STATE_CREATE_VNI_CONTEXT);
-
-    struct fi_info *prov_use = MPIDI_OFI_global.prov_use;
-
-    /* Each VNI context consists of domain, av, cq, cntr, etc.
-     *
-     * If MPIDI_OFI_VNI_USE_DOMAIN is true, each context is a separate domain,
-     * within which are each separate av, cq, cntr, ..., everything. Within the
-     * VNI context, it still can use either simple endpoint or scalable endpoint.
-     *
-     * If MPIDI_OFI_VNI_USE_DOMAIN is false, then all the VNI contexts will share
-     * the same domain and av, and use a single scalable endpoint. Separate VNI
-     * context will have its separate cq and separate tx and rx with the SEP.
-     *
-     * To accomodate both configurations, each context structure will have all fields
-     * including domain, av, cq, ... For "VNI_USE_DOMAIN", they are not shared.
-     * When not "VNI_USE_DOMAIN" or "VNI_USE_SEPCTX", domain, av, and ep are shared
-     * with the root (or 0th) VNI context.
-     */
-    struct fid_domain *domain;
-    struct fid_av *av;
-    struct fid_cntr *rma_cmpl_cntr;
-    struct fid_cq *cq;
-
-    struct fid_ep *ep;
-    struct fid_ep *tx;
-    struct fid_ep *rx;
-
-#ifdef MPIDI_OFI_VNI_USE_DOMAIN
-    mpi_errno = create_vni_domain(&domain, &av, &rma_cmpl_cntr);
-    MPIR_ERR_CHECK(mpi_errno);
-    mpi_errno = create_cq(domain, &cq);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    if (MPIDI_OFI_ENABLE_SCALABLE_ENDPOINTS) {
-        MPIDI_OFI_CALL(fi_scalable_ep(domain, prov_use, &ep, NULL), ep);
-        MPIDI_OFI_CALL(fi_scalable_ep_bind(ep, &av->fid, 0), bind);
-        MPIDI_OFI_CALL(fi_enable(ep), ep_enable);
-
-        mpi_errno = create_sep_tx(ep, 0, &tx, cq, rma_cmpl_cntr);
-        MPIR_ERR_CHECK(mpi_errno);
-        mpi_errno = create_sep_rx(ep, 0, &rx, cq, rma_cmpl_cntr);
-        MPIR_ERR_CHECK(mpi_errno);
-    } else {
-        MPIDI_OFI_CALL(fi_endpoint(domain, prov_use, &ep, NULL), ep);
-        MPIDI_OFI_CALL(fi_ep_bind(ep, &av->fid, 0), bind);
-        MPIDI_OFI_CALL(fi_ep_bind(ep, &cq->fid, FI_SEND | FI_RECV | FI_SELECTIVE_COMPLETION), bind);
-        MPIDI_OFI_CALL(fi_ep_bind(ep, &rma_cmpl_cntr->fid, FI_READ | FI_WRITE), bind);
-        MPIDI_OFI_CALL(fi_enable(ep), ep_enable);
-        tx = ep;
-        rx = ep;
-    }
-    MPIDI_OFI_global.ctx[vni].domain = domain;
-    MPIDI_OFI_global.ctx[vni].av = av;
-    MPIDI_OFI_global.ctx[vni].rma_cmpl_cntr = rma_cmpl_cntr;
-    MPIDI_OFI_global.ctx[vni].ep = ep;
-    MPIDI_OFI_global.ctx[vni].cq = cq;
-    MPIDI_OFI_global.ctx[vni].tx = tx;
-    MPIDI_OFI_global.ctx[vni].rx = rx;
-
-#else /* MPIDI_OFI_VNI_USE_SEPCTX */
-    if (vni == 0) {
-        mpi_errno = create_vni_domain(&domain, &av, &rma_cmpl_cntr);
-        MPIR_ERR_CHECK(mpi_errno);
-    } else {
-        domain = MPIDI_OFI_global.ctx[0].domain;
-        av = MPIDI_OFI_global.ctx[0].av;
-        rma_cmpl_cntr = MPIDI_OFI_global.ctx[0].rma_cmpl_cntr;
-    }
-    mpi_errno = create_cq(domain, &cq);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    if (vni == 0) {
-        if (MPIDI_OFI_ENABLE_SCALABLE_ENDPOINTS) {
-            MPIDI_OFI_CALL(fi_scalable_ep(domain, prov_use, &ep, NULL), ep);
-            MPIDI_OFI_CALL(fi_scalable_ep_bind(ep, &av->fid, 0), bind);
-            MPIDI_OFI_CALL(fi_enable(ep), ep_enable);
-        } else {
-            MPIDI_OFI_CALL(fi_endpoint(domain, prov_use, &ep, NULL), ep);
-            MPIDI_OFI_CALL(fi_ep_bind(ep, &av->fid, 0), bind);
-            MPIDI_OFI_CALL(fi_ep_bind(ep, &cq->fid, FI_SEND | FI_RECV | FI_SELECTIVE_COMPLETION),
-                           bind);
-            MPIDI_OFI_CALL(fi_ep_bind(ep, &rma_cmpl_cntr->fid, FI_READ | FI_WRITE), bind);
-            MPIDI_OFI_CALL(fi_enable(ep), ep_enable);
-        }
-    } else {
-        ep = MPIDI_OFI_global.ctx[0].ep;
-    }
-
-    if (MPIDI_OFI_ENABLE_SCALABLE_ENDPOINTS) {
-        mpi_errno = create_sep_tx(ep, vni, &tx, cq, rma_cmpl_cntr);
-        MPIR_ERR_CHECK(mpi_errno);
-        mpi_errno = create_sep_rx(ep, vni, &rx, cq, rma_cmpl_cntr);
-        MPIR_ERR_CHECK(mpi_errno);
-    } else {
-        tx = ep;
-        rx = ep;
-    }
-
-    if (vni == 0) {
-        MPIDI_OFI_global.ctx[0].domain = domain;
-        MPIDI_OFI_global.ctx[0].av = av;
-        MPIDI_OFI_global.ctx[0].rma_cmpl_cntr = rma_cmpl_cntr;
-        MPIDI_OFI_global.ctx[0].ep = ep;
-    }
-    MPIDI_OFI_global.ctx[vni].cq = cq;
-    MPIDI_OFI_global.ctx[vni].tx = tx;
-    MPIDI_OFI_global.ctx[vni].rx = rx;
-#endif
-
-    /* ------------------------------------------------------------------------ */
-    /* Construct:  Shared TX Context for RMA                                    */
-    /* ------------------------------------------------------------------------ */
-    if (vni == 0) {
-        mpi_errno = create_rma_stx_ctx(domain, &MPIDI_OFI_global.rma_stx_ctx);
-        MPIR_ERR_CHECK(mpi_errno);
-    }
-
-  fn_exit:
-    MPIR_FUNC_VERBOSE_EXIT(MPID_STATE_CREATE_VNI_CONTEXT);
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
 /* ---------------------------------------------------------- */
 /* Provider Selections and open_fabric()                      */
 /* ---------------------------------------------------------- */
@@ -1102,271 +682,6 @@ bool match_global_settings(struct fi_info *prov);
 static struct fi_info *pick_provider_from_list(const char *provname, struct fi_info *prov_list);
 static struct fi_info *pick_provider_by_name(const char *provname, struct fi_info *prov_list);
 static struct fi_info *pick_provider_by_global_settings(struct fi_info *prov_list);
-
-static int destroy_vni_context(int vni)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-#ifdef MPIDI_OFI_VNI_USE_DOMAIN
-    if (MPIDI_OFI_ENABLE_SCALABLE_ENDPOINTS) {
-        MPIDI_OFI_CALL(fi_close((fid_t) MPIDI_OFI_global.ctx[vni].tx), epclose);
-        MPIDI_OFI_CALL(fi_close((fid_t) MPIDI_OFI_global.ctx[vni].rx), epclose);
-        MPIDI_OFI_CALL(fi_close((fid_t) MPIDI_OFI_global.ctx[vni].cq), cqclose);
-
-        MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].ep->fid), epclose);
-        MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].av->fid), avclose);
-        MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].rma_cmpl_cntr->fid), cntrclose);
-        MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].domain->fid), domainclose);
-    } else {    /* normal endpoint */
-        MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].ep->fid), epclose);
-        MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].cq->fid), cqclose);
-        MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].av->fid), avclose);
-        MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].rma_cmpl_cntr->fid), cntrclose);
-        MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].domain->fid), domainclose);
-    }
-
-#else /* MPIDI_OFI_VNI_USE_SEPCTX */
-    if (MPIDI_OFI_ENABLE_SCALABLE_ENDPOINTS) {
-        MPIDI_OFI_CALL(fi_close((fid_t) MPIDI_OFI_global.ctx[vni].tx), epclose);
-        MPIDI_OFI_CALL(fi_close((fid_t) MPIDI_OFI_global.ctx[vni].rx), epclose);
-        MPIDI_OFI_CALL(fi_close((fid_t) MPIDI_OFI_global.ctx[vni].cq), cqclose);
-        if (vni == 0) {
-            MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].ep->fid), epclose);
-            MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].av->fid), avclose);
-            MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].rma_cmpl_cntr->fid), cntrclose);
-            MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].domain->fid), domainclose);
-        }
-    } else {    /* normal endpoint */
-        MPIR_Assert(vni == 0);
-        MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].ep->fid), epclose);
-        MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].cq->fid), cqclose);
-        MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].av->fid), avclose);
-        MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].rma_cmpl_cntr->fid), cntrclose);
-        MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.ctx[vni].domain->fid), domainclose);
-    }
-#endif
-    if (vni == 0) {
-        /* Close RMA scalable EP. */
-        if (MPIDI_OFI_global.rma_sep) {
-            /* All transmit contexts on RMA must be closed. */
-            MPIR_Assert(utarray_len(MPIDI_OFI_global.rma_sep_idx_array) ==
-                        MPIDI_OFI_global.max_rma_sep_tx_cnt);
-            utarray_free(MPIDI_OFI_global.rma_sep_idx_array);
-            MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.rma_sep->fid), epclose);
-        }
-
-        if (MPIDI_OFI_global.rma_stx_ctx != NULL) {
-            MPIDI_OFI_CALL(fi_close(&MPIDI_OFI_global.rma_stx_ctx->fid), stx_ctx_close);
-        }
-    }
-
-  fn_exit:
-    MPIR_FUNC_VERBOSE_EXIT(MPID_STATE_DESTROY_VNI_CONTEXT);
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-static int create_vni_domain(struct fid_domain **p_domain, struct fid_av **p_av,
-                             struct fid_cntr **p_cntr)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    /* ---- domain ---- */
-    struct fid_domain *domain;
-    MPIDI_OFI_CALL(fi_domain(MPIDI_OFI_global.fabric, MPIDI_OFI_global.prov_use, &domain, NULL),
-                   opendomain);
-    *p_domain = domain;
-
-    /* ---- av ---- */
-    /* ----
-     * Attempt to open a shared address vector read-only.
-     * The open will fail if the address vector does not exist.
-     * Otherwise, set MPIDI_OFI_global.got_named_av and
-     * copy the map_addr.
-     */
-    if (try_open_shared_av(domain, p_av)) {
-        MPIDI_OFI_global.got_named_av = 1;
-    }
-
-    if (!MPIDI_OFI_global.got_named_av) {
-        struct fi_av_attr av_attr;
-        memset(&av_attr, 0, sizeof(av_attr));
-        if (MPIDI_OFI_ENABLE_AV_TABLE) {
-            av_attr.type = FI_AV_TABLE;
-        } else {
-            av_attr.type = FI_AV_MAP;
-        }
-        av_attr.rx_ctx_bits = MPIDI_OFI_MAX_ENDPOINTS_BITS;
-
-        av_attr.name = NULL;
-        av_attr.flags = 0;
-        MPIDI_OFI_CALL(fi_av_open(domain, &av_attr, p_av, NULL), avopen);
-    }
-
-    /* ---- other sharable objects ---- */
-    struct fi_cntr_attr cntr_attr;
-    memset(&cntr_attr, 0, sizeof(cntr_attr));
-    cntr_attr.events = FI_CNTR_EVENTS_COMP;
-    cntr_attr.wait_obj = FI_WAIT_UNSPEC;
-    MPIDI_OFI_CALL(fi_cntr_open(domain, &cntr_attr, p_cntr, NULL), openct);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-static int create_cq(struct fid_domain *domain, struct fid_cq **p_cq)
-{
-    int mpi_errno = MPI_SUCCESS;
-    struct fi_cq_attr cq_attr;
-    memset(&cq_attr, 0, sizeof(cq_attr));
-    cq_attr.format = FI_CQ_FORMAT_TAGGED;
-    MPIDI_OFI_CALL(fi_cq_open(domain, &cq_attr, p_cq, NULL), opencq);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-static int create_sep_tx(struct fid_ep *ep, int idx, struct fid_ep **p_tx,
-                         struct fid_cq *cq, struct fid_cntr *cntr)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    struct fi_tx_attr tx_attr;
-    tx_attr = *(MPIDI_OFI_global.prov_use->tx_attr);
-    tx_attr.op_flags = FI_COMPLETION;
-    if (MPIDI_OFI_ENABLE_RMA || MPIDI_OFI_ENABLE_ATOMICS)
-        tx_attr.op_flags |= FI_DELIVERY_COMPLETE;
-    tx_attr.caps = 0;
-
-    if (MPIDI_OFI_ENABLE_TAGGED)
-        tx_attr.caps = FI_TAGGED;
-
-    /* RMA */
-    if (MPIDI_OFI_ENABLE_RMA)
-        tx_attr.caps |= FI_RMA;
-    if (MPIDI_OFI_ENABLE_ATOMICS)
-        tx_attr.caps |= FI_ATOMICS;
-    /* MSG */
-    tx_attr.caps |= FI_MSG;
-    tx_attr.caps |= FI_NAMED_RX_CTX;    /* Required for scalable endpoints indexing */
-
-    MPIDI_OFI_CALL(fi_tx_context(ep, idx, &tx_attr, p_tx, NULL), ep);
-    MPIDI_OFI_CALL(fi_ep_bind(*p_tx, &cq->fid, FI_SEND | FI_SELECTIVE_COMPLETION), bind);
-    MPIDI_OFI_CALL(fi_ep_bind(*p_tx, &cntr->fid, FI_WRITE | FI_READ), bind);
-    MPIDI_OFI_CALL(fi_enable(*p_tx), ep_enable);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-static int create_sep_rx(struct fid_ep *ep, int idx, struct fid_ep **p_rx,
-                         struct fid_cq *cq, struct fid_cntr *cntr)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    struct fi_rx_attr rx_attr;
-    rx_attr = *(MPIDI_OFI_global.prov_use->rx_attr);
-    rx_attr.caps = 0;
-
-    if (MPIDI_OFI_ENABLE_TAGGED) {
-        rx_attr.caps |= FI_TAGGED;
-        rx_attr.caps |= FI_DIRECTED_RECV;
-    }
-
-    if (MPIDI_OFI_ENABLE_RMA)
-        rx_attr.caps |= FI_RMA | FI_REMOTE_READ | FI_REMOTE_WRITE;
-    if (MPIDI_OFI_ENABLE_ATOMICS)
-        rx_attr.caps |= FI_ATOMICS;
-    rx_attr.caps |= FI_MSG;
-    rx_attr.caps |= FI_MULTI_RECV;
-    rx_attr.caps |= FI_NAMED_RX_CTX;    /* Required for scalable endpoints indexing */
-
-    MPIDI_OFI_CALL(fi_rx_context(ep, idx, &rx_attr, p_rx, NULL), ep);
-    MPIDI_OFI_CALL(fi_ep_bind(*p_rx, &cq->fid, FI_RECV), bind);
-    MPIDI_OFI_CALL(fi_enable(*p_rx), ep_enable);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-static int try_open_shared_av(struct fid_domain *domain, struct fid_av **p_av)
-{
-#ifdef MPIDI_OFI_VNI_USE_DOMAIN
-    /* shared/named av table cannot be used when multiple fi_domain is enabled */
-    return 0;
-#else
-    struct fi_av_attr av_attr;
-    memset(&av_attr, 0, sizeof(av_attr));
-    if (MPIDI_OFI_ENABLE_AV_TABLE) {
-        av_attr.type = FI_AV_TABLE;
-    } else {
-        av_attr.type = FI_AV_MAP;
-    }
-    av_attr.rx_ctx_bits = MPIDI_OFI_MAX_ENDPOINTS_BITS;
-
-    char av_name[128];
-    MPL_snprintf(av_name, sizeof(av_name), "FI_NAMED_AV_%d\n", MPIR_Process.appnum);
-    av_attr.name = av_name;
-    av_attr.flags = FI_READ;
-    av_attr.map_addr = 0;
-
-    if (0 == fi_av_open(domain, &av_attr, p_av, NULL)) {
-        /* TODO - the copy from the pre-existing av map into the 'MPIDI_OFI_AV' */
-        /* is wasteful and should be changed so that the 'MPIDI_OFI_AV' object  */
-        /* directly references the mapped fi_addr_t array instead               */
-        fi_addr_t *mapped_table = (fi_addr_t *) av_attr.map_addr;
-        for (int i = 0; i < MPIR_Process.size; i++) {
-            MPIDI_OFI_AV(&MPIDIU_get_av(0, i)).dest = mapped_table[i];
-#if MPIDI_OFI_ENABLE_ENDPOINTS_BITS
-            MPIDI_OFI_AV(&MPIDIU_get_av(0, i)).ep_idx = 0;
-#endif
-            MPL_DBG_MSG_FMT(MPIDI_CH4_DBG_MAP, VERBOSE,
-                            (MPL_DBG_FDEST, " grank mapped to: rank=%d, av=%p, dest=%" PRIu64,
-                             i, (void *) &MPIDIU_get_av(0, i), mapped_table[i]));
-        }
-        return 1;
-    } else {
-        return 0;
-    }
-#endif
-}
-
-static int create_rma_stx_ctx(struct fid_domain *domain, struct fid_stx **p_rma_stx_ctx)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    if (MPIDI_OFI_ENABLE_SHARED_CONTEXTS) {
-        int ret;
-        struct fi_tx_attr tx_attr;
-        memset(&tx_attr, 0, sizeof(tx_attr));
-        /* A shared transmit context’s attributes must be a union of all associated
-         * endpoints' transmit capabilities. */
-        tx_attr.caps = FI_RMA | FI_WRITE | FI_READ | FI_ATOMIC;
-        tx_attr.msg_order = FI_ORDER_RAR | FI_ORDER_RAW | FI_ORDER_WAR | FI_ORDER_WAW;
-        tx_attr.op_flags = FI_DELIVERY_COMPLETE | FI_COMPLETION;
-        MPIDI_OFI_CALL_RETURN(fi_stx_context(domain, &tx_attr, p_rma_stx_ctx, NULL), ret);
-        if (ret < 0) {
-            MPL_DBG_MSG(MPIDI_CH4_DBG_GENERAL, VERBOSE,
-                        "Failed to create shared TX context for RMA, "
-                        "falling back to global EP/counter scheme");
-            *p_rma_stx_ctx = NULL;
-        }
-    }
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
 
 static int open_fabric(void)
 {
